Create pcpm-server-v1.py
Version 1.0 

This is only "Zavier Chambers" code and is not the FINAL product. 

I wanted to layout my thinking and display something to following quickly for my clients in class. I will stop by the computer lab for a test to see if I can run this script. 

Thanks everyone!